+++ docs/make.jl
@@ -1,7 +1,6 @@
 using Documenter, Microbiome
 
 makedocs(
-    format = :html,
     sitename = "Microbiome.jl",
     pages = [
         "Home" => "index.md",
@@ -9,7 +8,9 @@ makedocs(
         "Distances & Dissimilarity" => "distances.md",
         "Contributing" => "contributing.md"
     ],
-    authors = "Kevin Bonham, PhD"
+    authors = "Kevin Bonham, PhD",
+    format = Documenter.HTML(
+        prettyurls = get(ENV, "CI", nothing) == "true")
 )
 
 deploydocs(