+++ src/Microbiome.jl
@@ -1,10 +1,6 @@
 module Microbiome
 
 export
-    # Types
-    DistanceMatrix,
-    PCoA,
-
     # Functions
     ## abundance
     abundancetable,
@@ -23,15 +19,7 @@ export
     getsample,
     samplenames,
     sampletotals,
-    ## distances
-    getdm,
-    getrowdm,
-    pcoa,
-    eigenvalue,
-    principalcoord,
-    variance,
-    optimalorder,
-    optimalorder!,
+    ## Diversity
     ginisimpson,
     shannon,
     present,
@@ -40,13 +28,10 @@ export
 
 using Reexport
 @reexport using SpatialEcology
-@reexport using Distances
 
-using LinearAlgebra
 using Statistics
 using StatsBase
 using DataFrames
-using Clustering
 
 import SpatialEcology: @forward_func
 import Base: getindex, setindex, length
@@ -54,6 +39,5 @@ import Base: getindex, setindex, length
 include("ecotranslations.jl")
 include("abundances.jl")
 include("distances.jl")
-include("leafordering.jl")
 
 end  # module Microbiome