+++ test/runtests.jl
@@ -1,8 +1,5 @@
 using Microbiome
-using Distances
-using LinearAlgebra
 using DataFrames
-using Clustering
 using Random
 using Test
 
@@ -82,37 +79,6 @@ end
         M, ["sample_$x" for x in 1:10],
         ["feature_$x" for x in 1:100])
 
-    dm1 = getdm(M, BrayCurtis())
-    dm2 = getdm(df, BrayCurtis())
-    dm = getdm(abund, BrayCurtis())
-
-    @test dm.dm == dm1.dm == dm2.dm
-
-    rowdm1 = getrowdm(M, BrayCurtis())
-    rowdm2 = getrowdm(df, BrayCurtis())
-    rowdm = getrowdm(abund, BrayCurtis())
-
-    @test rowdm.dm == rowdm1.dm == rowdm2.dm
-
-    @test size(dm) == (10, 10)
-    @test size(rowdm) == (100, 100)
-    for i in 1:10; @test dm[i,i] == 0; end
-    for i in 1:100; @test rowdm[i,i] == 0; end
-
-    # PCoA
-    p = pcoa(dm, correct_neg=true)
-    @test sum(p.variance_explained) ≈ 1
-    for i in 1:size(p, 2)
-        @test eigenvalue(p, i) > 0
-        @test typeof(eigenvalue(p, i)) <: Real
-    end
-
-    @test sum([variance(p, i) for i in 1:size(p,2)]) ≈ 1
-    @test sort(variance(p, 1:size(p,2)), rev=true) == variance(p, 1:size(p,2))
-
-    @test length(principalcoord(p, 1)) == size(dm, 1)
-    @test principalcoord(p, 1:size(p,2)) == p.eigenvectors
-
     # Diversity indicies
     R = 100
     s1 = rand(R) # high diversity
@@ -127,17 +93,7 @@ end
     @test ginisimpson(s1) > ginisimpson(s2)
     @test ginisimpson(s3) ≈ 1. - 1/R
     @test ginisimpson(s4) ≈ 0.
-end
 
-@testset "Leaf Ordering" begin
-    Random.seed!(42)
-    m = rand(100, 10)
-
-    dm = pairwise(BrayCurtis(), m, dims=2)
-    h = hclust(dm, linkage=:single);
-
-    ordered = optimalorder(h, dm)
-
-    @test ordered.order == [7, 3, 1, 9, 2, 6, 10, 4, 5, 8]
-    @test ordered.merges == h.merges
+    @test length(shannon(abund)) == 10
+    @test length(ginisimpson(abund)) == 10
 end