+++ NEWS.md
@@ -1,5 +1,13 @@
 # News for Microbiome.jl
 
+## v0.5.0
+
+Major Changes:  
+- Dropped Microbiome.jl-defined types and methods for DistanceMatrix. Use Distances.jl instead. [Requires this SpatialEcology PR](https://github.com/EcoJulia/SpatialEcology.jl/pull/36)
+- Dropped Microbiome.jl-defined types and methods for PCoA. Use MDS from MultivariateStats.jl instead. [Requires this MultivariateStats PR](https://github.com/JuliaStats/MultivariateStats.jl/pull/85)
+- Dropped Hclust leaf ordering. Added to Clustering.jl instead ([see Clustering.jl PR](https://github.com/JuliaStats/Clustering.jl/pull/170))
+
+
 ## v0.4.1
 
 Major Changes: